requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic[email]>=2.9.0
email-validator>=2.1.0
requests>=2.32.3
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
